requirements.txt
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.0
PyJWT==2.6.0
six==1.16.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
validators==0.20.0
Werkzeug==2.2.3
